+++ packages/connectors/src/exports/index.ts
@@ -6,6 +6,8 @@ export {
   mock,
 } from '@wagmi/core'
 
+export { type BaseAccountParameters, baseAccount } from '../baseAccount.js'
+
 export {
   type CoinbaseWalletParameters,
   coinbaseWallet,
@@ -14,10 +16,8 @@ export {
 export { type MetaMaskParameters, metaMask } from '../metaMask.js'
 
 export { type SafeParameters, safe } from '../safe.js'
-
+export { version } from '../version.js'
 export {
   type WalletConnectParameters,
   walletConnect,
 } from '../walletConnect.js'
-
-export { version } from '../version.js'