+++ packages/core/src/actions/watchBlocks.ts
@@ -1,10 +1,9 @@
+import type { BlockTag, Chain, Transport, WebSocketTransport } from 'viem'
 import {
   type WatchBlocksParameters as viem_WatchBlocksParameters,
   type WatchBlocksReturnType as viem_WatchBlocksReturnType,
   watchBlocks as viem_watchBlocks,
 } from 'viem/actions'
-
-import type { BlockTag, Chain, Transport, WebSocketTransport } from 'viem'
 import type { Config } from '../createConfig.js'
 import type { SelectChains } from '../types/chain.js'
 import type {