+++ packages/core/src/connectors/mock.test.ts
@@ -23,14 +23,16 @@ test('setup', () => {
   expectTypeOf<ConnectFnParameters['foo']>().toMatchTypeOf<string | undefined>()
 })
 
-test('behavior: features.connectError', () => {
+test('behavior: features.connectError', async () => {
   const connectorFn = mock({ accounts, features: { connectError: true } })
   const connector = config._internal.connectors.setup(connectorFn)
-  expect(() => connector.connect()).rejects.toThrowErrorMatchingInlineSnapshot(`
+  await expect(() =>
+    connector.connect(),
+  ).rejects.toThrowErrorMatchingInlineSnapshot(`
     [UserRejectedRequestError: User rejected the request.
 
     Details: Failed to connect.
-    Version: viem@2.29.2]
+    Version: viem@2.31.7]
   `)
 })
 
@@ -49,7 +51,7 @@ test('behavior: connector.getProvider request errors', async () => {
   ) as ReturnType<typeof connectorFn>
   const provider = await connector.getProvider()
 
-  expect(
+  await expect(
     provider.request({
       method: 'eth_signTypedData_v4',
       params: [] as any,
@@ -58,10 +60,10 @@ test('behavior: connector.getProvider request errors', async () => {
     [UserRejectedRequestError: User rejected the request.
 
     Details: Failed to sign typed data.
-    Version: viem@2.29.2]
+    Version: viem@2.31.7]
   `)
 
-  expect(
+  await expect(
     provider.request({
       method: 'wallet_switchEthereumChain',
       params: [] as any,
@@ -70,10 +72,10 @@ test('behavior: connector.getProvider request errors', async () => {
     [UserRejectedRequestError: User rejected the request.
 
     Details: Failed to switch chain.
-    Version: viem@2.29.2]
+    Version: viem@2.31.7]
   `)
 
-  expect(
+  await expect(
     provider.request({
       method: 'wallet_watchAsset',
       params: [] as any,
@@ -82,10 +84,10 @@ test('behavior: connector.getProvider request errors', async () => {
     [UserRejectedRequestError: User rejected the request.
 
     Details: Failed to switch chain.
-    Version: viem@2.29.2]
+    Version: viem@2.31.7]
   `)
 
-  expect(
+  await expect(
     provider.request({
       method: 'personal_sign',
       params: [] as any,
@@ -94,7 +96,7 @@ test('behavior: connector.getProvider request errors', async () => {
     [UserRejectedRequestError: User rejected the request.
 
     Details: Failed to sign message.
-    Version: viem@2.29.2]
+    Version: viem@2.31.7]
   `)
 })