+++ packages/react/src/hooks/useTransactionReceipt.ts
@@ -6,13 +6,13 @@ import type {
   ResolvedRegister,
 } from '@wagmi/core'
 import type { Compute } from '@wagmi/core/internal'
+import type { GetTransactionReceiptQueryFnData } from '@wagmi/core/query'
 import {
   type GetTransactionReceiptData,
   type GetTransactionReceiptOptions,
   type GetTransactionReceiptQueryKey,
   getTransactionReceiptQueryOptions,
 } from '@wagmi/core/query'
-import type { GetTransactionReceiptQueryFnData } from '@wagmi/core/query'
 import type { ConfigParameter, QueryParameter } from '../types/properties.js'
 import { type UseQueryReturnType, useQuery } from '../utils/query.js'
 import { useChainId } from './useChainId.js'