+++ packages/react/src/hooks/useVerifyTypedData.test-d.ts
@@ -1,7 +1,6 @@
 import type { typedData } from '@wagmi/test'
-import { expectTypeOf, test } from 'vitest'
-
 import type { Address } from 'viem'
+import { expectTypeOf, test } from 'vitest'
 import {
   type UseVerifyTypedDataParameters,
   useVerifyTypedData,