+++ packages/react/src/hooks/useWatchContractEvent.test-d.ts
@@ -1,4 +1,4 @@
-import { http, createConfig, webSocket } from '@wagmi/core'
+import { createConfig, http, webSocket } from '@wagmi/core'
 import { mainnet, optimism } from '@wagmi/core/chains'
 import { abi } from '@wagmi/test'
 import { expectTypeOf, test } from 'vitest'
@@ -42,8 +42,6 @@ test('behavior: no eventName', () => {
     onLogs(logs) {
       expectTypeOf(logs[0]!.eventName).toEqualTypeOf<'Transfer' | 'Approval'>()
       expectTypeOf(logs[0]!.args).toEqualTypeOf<
-        | Record<string, unknown>
-        | readonly unknown[]
         | {
             from?: `0x${string}` | undefined
             to?: `0x${string}` | undefined