+++ packages/vue/src/composables/useChainId.ts
@@ -1,11 +1,11 @@
 import {
   type Config,
   type GetChainIdReturnType,
-  type ResolvedRegister,
   getChainId,
+  type ResolvedRegister,
   watchChainId,
 } from '@wagmi/core'
-import { type Ref, onScopeDispose, readonly, ref } from 'vue'
+import { onScopeDispose, type Ref, readonly, ref } from 'vue'
 
 import type { ConfigParameter } from '../types/properties.js'
 import { useConfig } from './useConfig.js'