+++ site/snippets/vue/config-chain-properties.ts
@@ -1,4 +1,4 @@
-import { http, createConfig } from '@wagmi/vue'
+import { createConfig, http } from '@wagmi/vue'
 import { base, celo, mainnet } from '@wagmi/vue/chains'
 
 export const config = createConfig({